CRUD de usuarios funcionando solo con TOKEN, excepto CREATE

+++ src/main/java/com/project/user/controllers/rest/UserController.java
@@ -52,28 +52,35 @@ public class UserController {
         return userDao.getUsers();
     }
 
+    // * Update
+    // Actualiza el registro si el usuario está logueado
+    @RequestMapping(value = "api/users", method = RequestMethod.PUT)
+    public String updateUser(@RequestBody userModel user, @RequestHeader(value = "Authorization") String token) {
+
+        if (!validateToken(token)) {
+            return null;
+        }
+        userDao.updateUser(user);
+        return "OK";
+    }
+
+    // * Delete
+    // Elimina un registro si el usuario está logueado
+    @RequestMapping(value = "api/users/{id}", method = RequestMethod.DELETE)
+    public void deleteUsers(@PathVariable Long id, @RequestHeader(value = "Authorization") String token) {
+        if (!validateToken(token)) {
+            System.out.println("It's not possible to delete user");
+        } else {
+            userDao.deleteUsers(id);
+        }
+    }
+
     @RequestMapping(value = "api/users/{id}", method = RequestMethod.GET)
     public List<userModel> getUser(@PathVariable Long id) {
         return userDao.getUser(id);
     }
 
-    @RequestMapping(value = "api/users/{id}", method = RequestMethod.DELETE)
-    public void deleteUsers(@PathVariable Long id) {
-        userDao.deleteUsers(id);
-    }
-
-    @RequestMapping(value = "api/users", method = RequestMethod.PUT)
-    public String updateUser(@RequestBody userModel user) {
-        userDao.updateUser(user);
-        return "OK";
-        // if (userDao.verifyUser(user)) {
-        // userDao.updateUser(user);
-        // return "OK";
-        // } else {
-        // return "FAIL";
-        // }
-    }
-
+    // Valida que el token sea el mismo del usuario logeado
     private boolean validateToken(String token) {
         String userID = jwtUtil.getKey(token);
         return userID != null;
@@ -84,12 +91,12 @@ public class UserController {
      * public List<userModel> getUsers(){
      * return UserDao.getUsers();
      * }
-     * 
+     *
      */
 
     /*
      * Funciona, devuelve la lista
-     * 
+     *
      * @RequestMapping(value = "api/usuario123", method = RequestMethod.GET)
      * public List<userModel> getUsuario(){
      * List<userModel> users = new ArrayList<>();
@@ -100,7 +107,7 @@ public class UserController {
      * usuario.setEmail("alexmata@hotmail.com");
      * usuario.setPassword("1234567");
      * usuario.setType("admin");
-     * 
+     *
      * userModel usuario2 = new userModel();
      * usuario2.setId(123L);
      * usuario2.setName("Alex2");
@@ -108,7 +115,7 @@ public class UserController {
      * usuario2.setEmail("alexmata2@hotmail.com");
      * usuario2.setPassword("12345678");
      * usuario2.setType("noAdmin");
-     * 
+     *
      * users.add(usuario);
      * users.add(usuario2);
      * return users;
